Separação das funções
Separei as funções para deixar o código principal mais organizado, facilitando a implementação futura.
