+++ todo-cli/src/adicionar.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include <string>
+#include <iostream>
+
+using namespace std;
+
+void adicionar_tarefa(const string & tarefa){
+    cout << "Seu arquivo foi adicionado: " << tarefa << endl;
+/*Mostrará para o usuário que sua tarefa foi adicionada com sucesso;
+Para adicionar a lógica para salvar a tarefa em um arquivo ou banco de dados. Retirei do código principal para permitir melhor organização.*/
+}
