+++ todo-cli/src/remover.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include <string>
+#include <iostream>
+
+using namespace std;
+
+void remover_tarefa(const string & tarefa){
+    cout << "Removendo tarefa: " << tarefa << endl;
+
+/*Implementar a lógica para remover a tarefa de um arquivo ou banco de dados. Entretanto, ainda não está configurada e ainda não foi inserida no objeto principal da interface*/
+}
